Fixed insane duplicate emission problem with externs
fixed gitignore in tests dir

+++ run_test.py
@@ -136,8 +136,8 @@ def main():
             print(f"\n\033[1;31m{test}:\033[0m")
             print(out.strip() or "<no output>")
 
-    for fname in os.listdir("."):
-        if not fname.endswith(".ela"):
+        for fname in os.listdir("."):
+        if not fname.endswith(".ela") and not fname.endswith(".gitignore"):
             try:
                 os.remove(fname)
             except Exception as e:
